restructure code - seperate extractor from builder
