--- lib/qippet/box/text_box.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-require_relative "./box"
-
-module Qippet
-  class TextBox < Box
-  end
-end
