+++ lib/qippet/extract.rb
@@ -1,30 +1,20 @@
 # frozen_string_literal: true
 
 module Qippet
-  # Extraction Module
-  module Extraction
-    extend ActiveSupport::Autoload
+  # Extract Class
+  class Extract
+    class << self
+      def from_file(path)
+        path = Pathname.new(Dir.pwd).join(path)
+        read_file(path)
+      end
 
-    autoload :Xml
+      private
 
-    # Extract Class
-    class Extract
-      include Xml
-
-      class << self
-        def from_file(path)
-          path = Pathname.new(Dir.pwd).join(path)
-          content = read_file(path)
-          return nil if content.nil?
-
-          extract_from_xml_content(content)
-        end
-
-        private
-
-        def read_file(path)
-          File.read(path)
-        end
+      def read_file(path)
+        File.read(path)
+      rescue Errno::ENOENT, Errno::EISDIR
+        nil
       end
     end
   end
