+++ spec/qippet_spec.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require "qippet"
+require "qippet/builder/xml"
 
 RSpec.describe Qippet do
   it "has a version number" do
@@ -21,16 +22,12 @@ RSpec.describe Qippet do
     expect { Qippet.generate("", output_file) }.to output("File does not contain a valid Qippet structure\n").to_stdout
   end
 
-  it "should output error if file does not exist" do
-    expect { Qippet.generate("random.rs", "") }.to raise_error(Errno::ENOENT)
-  end
-
   it "Should call extract from xml if file exists" do
-    allow(Qippet::Extraction::Extract).to receive(:extract_from_xml_content).and_call_original
+    allow(Qippet::Builder::Xml).to receive(:build).and_call_original
     output_file = "image.png"
     file_path = "spec/assets/pop.xml"
     Qippet.generate(file_path, output_file)
-    expect(Qippet::Extraction::Extract).to have_received(:extract_from_xml_content)
+    expect(Qippet::Builder::Xml).to have_received(:build)
   end
 end
